[throttle] fix: a compile error that only the stable Xcode could see
With the plugin-validation flags in place, the hardened Release build finally ran
on CI for the first time since 3.2.96 — and failed to compile.
`data.split(separator: 0x0A)` is ambiguous on `Data` under the stable toolchain:
the integer literal leaves the element type to inference and two `split`
overloads match. Xcode-beta resolves it, so it built here every time.

This is the second thing the red CI was hiding, and it is the more interesting
one: this Mac has only Xcode-beta, so local success is not evidence the shipped
toolchain agrees. The pinned SwiftLint fixed the same class of problem for
linting; this is the compiler equivalent, and CI is the only place it shows.

+++ Throttle/Services/MissionRuntimeService.swift
@@ -616,7 +616,7 @@ enum MissionRuntimeService {
         guard let handle = try? FileHandle(forReadingFrom: url) else { return nil }
         defer { try? handle.close() }
         let data = (try? handle.read(upToCount: byteLimit)) ?? Data()
-        return data.split(separator: 0x0A).map(Data.init)
+        return data.split(separator: UInt8(0x0A)).map(Data.init)
     }
 
     nonisolated private static func tailJSONLines(url: URL, byteLimit: UInt64) -> [Data]? {
@@ -626,7 +626,7 @@ enum MissionRuntimeService {
         let start = end > byteLimit ? end - byteLimit : 0
         guard (try? handle.seek(toOffset: start)) != nil,
               let data = try? handle.readToEnd() else { return nil }
-        var lines = data.split(separator: 0x0A).map(Data.init)
+        var lines = data.split(separator: UInt8(0x0A)).map(Data.init)
         if start > 0, !lines.isEmpty { lines.removeFirst() }
         return lines
     }
